// src/types/github.ts
export type RepoMeta = {
	id: number;
	name: string;
	full_name: string;
	html_url: string;
	description?: string | null;
	private?: boolean;
	visibility?: string;
	fork?: boolean;
	language?: string | null;
	topics?: string[];
	stargazers_count?: number;
	forks_count?: number;
	open_issues_count?: number;
	watchers_count?: number;
	created_at?: string;
	updated_at?: string;
	pushed_at?: string;
	license?: { name?: string; spdx_id?: string } | null;
	owner?: { login?: string; html_url?: string; avatar_url?: string } | null;
	_score?: number;
};